requirements.txt
aiohttp>=3.8.0
pydantic>=1.10.0
python-dateutil>=2.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
python-dotenv